refactor(reports): address PR review + decompose ReportPage
Addresses every reviewable finding on PR #2516 (CodeRabbit, Copilot,
React Doctor) plus the giant-component warning on the route file's
top-level component, which the original refactor left as a follow-up.

CodeRabbit / Copilot inline review
----------------------------------
1. **overdue-items-content.tsx:4** — Copilot caught the file header
   labelling Overdue Items as "(R3)" when the registry marks it as R6.
   Fixed.

2. **idle-assets-content.tsx / overdue-items-content.tsx** — Per-row
   `valuation` / `valueAtRisk` cell renderers used truthiness checks,
   so a real `0` rendered as "—". Switched the table cells to `!= null`
   so a legitimate $0 valuation now displays as "$0". Hero summary
   aggregates were *not* changed: those values come from
   `(kpis.find(...).rawValue as number) || 0`, so they can never be
   null at the display site, and `> 0` correctly indicates "no data".

3. **idle-threshold-selector.tsx** — Segmented control was visually
   stateful but had no toggle semantics for assistive tech. Added
   `aria-pressed` per button and `role="group"` + `aria-labelledby`
   on the wrapper (referencing the existing "Unused for:" label). Per
   WCAG 2.1 AA + the project's accessibility rule.

4. **monthly-booking-trends-content.tsx** — The trend explanation
   lived in a hover-only tooltip on a non-focusable `<div title=...>`,
   so keyboard and touch users couldn't reach `trendDescription`.
   Replaced with the shared `InfoTooltip` (sibling reports already
   use it) so the explanation is reachable via keyboard focus.

5. **timeframe-range-indicator.tsx** — Used `toLocaleDateString()`
   directly, violating the project rule that all UI dates must go
   through the shared `DateS` component (which respects user locale +
   timezone hints). Reworked to render both endpoints via `DateS`.
   Year-show heuristic preserved.

React Doctor (5x no-usememo-simple-expression)
----------------------------------------------
Each `*-content.tsx` had `const columns = useMemo(() => CONST, [])`
where `CONST` is already at module scope. The useMemo wrapper added
zero stability — the const reference is already stable across every
render — so it was pure scaffolding that misled future readers into
thinking the memo was load-bearing. Replaced with `const columns =
CONST;` (and dropped the now-unused `useMemo` import) in idle-assets,
asset-utilization, asset-activity, asset-inventory, top-booked-assets.

The custody-snapshot file is unaffected — its columns close over
`maxDays` from props, so the inline `useMemo([..., [maxDays])` is
genuinely necessary there.

React Doctor (no-giant-component on ReportPage)
-----------------------------------------------
The original refactor extracted the inline Content components beneath
ReportPage but left ReportPage itself at 413 lines (8 useCallback
handlers, a 50-line export flow, a 130-line dispatch switch, plus
page chrome JSX). React Doctor's giant-component warning fired on
this. Decomposed ReportPage into 5 new modules under
`~/components/reports/`:

  - `use-report-row-handlers.ts` — collapses the 8 per-report
    `useCallback`s into a single hook returning two stable handlers
    (`onBookingRowClick`, `onAssetRowClick`). TypeScript's structural
    function-type compatibility means `(row: { bookingId: string })
    => void` satisfies every per-report `onRowClick` prop.

  - `use-csv-export.ts` — owns the `isExporting` state plus the full
    fetch → blob → download → toast flow. Returns `{ isExporting,
    handleExport }`.

  - `report-export-actions.tsx` — Header children: the conditional
    PDF button + the CSV button.

  - `report-filter-bar.tsx` — owns its own `useSearchParams` and
    renders whichever filter row applies (`TimeframePicker` +
    `TimeframeRangeIndicator`, or `IdleThresholdSelector`). Absorbs
    the `showTimeframePicker` and `getExcludedPresets` predicates.

  - `report-content-switch.tsx` — the 130-line dispatch switch. Also
    absorbs the empty-state branch and the three
    `getEmptyStateTitle/Description/Cta` helpers, so ReportPage no
    longer needs to know about them.

After this, ReportPage is 85 lines: hooks + a clean JSX tree of
named composition pieces. The route file is 359 lines (down from 836,
which itself was down from the original 2832 in the prior commit).

Verification
------------
`pnpm webapp:validate` green: typecheck, lint, prettier, 1777 tests.
`pnpm exec react-doctor . --diff` reports 100/100 on the changed
files (down from 99/100 with one warning, down from the initial 6
warnings on the original PR).
