+++ apps/webapp/app/components/reports/timeframe-range-indicator.tsx
@@ -15,6 +15,7 @@
  * @see {@link file://./timeframe-picker.tsx}
  */
 
+import { DateS } from "~/components/shared/date";
 import type { ResolvedTimeframe } from "~/modules/reports/types";
 
 /** Props for {@link TimeframeRangeIndicator}. */
@@ -27,6 +28,10 @@ type Props = {
  * Shows the actual date range and day count for the selected timeframe.
  * Helps users understand why "This quarter" might show fewer bookings than "Last 90 days"
  * (e.g., Q2 starting Apr 1 is only 24 days so far vs a full 90 days).
+ *
+ * Dates render through the shared {@link DateS} component so they respect
+ * the user's locale + timezone hints (per the project-wide rule that all
+ * UI dates must go through `DateS`).
  */
 export function TimeframeRangeIndicator({ timeframe }: Props) {
   const { from, to } = timeframe;
@@ -44,26 +49,29 @@ export function TimeframeRangeIndicator({ timeframe }: Props) {
       (toMidnight.getTime() - fromMidnight.getTime()) / (1000 * 60 * 60 * 24)
     ) + 1;
 
-  // Format dates compactly
-  const formatDate = (date: Date) =>
-    date.toLocaleDateString("en-US", {
-      month: "short",
-      day: "numeric",
-    });
-
-  // Check if dates span different years
+  // Decide whether to show the year on each endpoint. Mirrors the previous
+  // hand-rolled logic: hide year when both dates fall in the current year
+  // and don't span across years; otherwise show it where it disambiguates.
   const fromYear = from.getFullYear();
   const toYear = to.getFullYear();
   const showYear = fromYear !== toYear || toYear !== new Date().getFullYear();
-
-  const fromStr =
-    formatDate(from) + (showYear && fromYear !== toYear ? `, ${fromYear}` : "");
-  const toStr = formatDate(to) + (showYear ? `, ${toYear}` : "");
+  const fromOptions: Intl.DateTimeFormatOptions = {
+    month: "short",
+    day: "numeric",
+    ...(showYear && fromYear !== toYear ? { year: "numeric" } : {}),
+  };
+  const toOptions: Intl.DateTimeFormatOptions = {
+    month: "short",
+    day: "numeric",
+    ...(showYear ? { year: "numeric" } : {}),
+  };
 
   return (
     <span className="hidden text-xs text-gray-500 sm:inline-flex sm:items-center sm:gap-1.5">
-      <span>
-        {fromStr} – {toStr}
+      <span className="inline-flex items-center gap-1">
+        <DateS date={from} options={fromOptions} />
+        <span>–</span>
+        <DateS date={to} options={toOptions} />
       </span>
       <span className="text-gray-300">·</span>
       <span className="font-medium text-gray-600">
